+++ movie.php
@@ -126,8 +126,10 @@
                                 <th>內容簡介</th>
                                 <th>詳細</th>
                                 <?php
-                                    if($_SESSION['mId']=='v123456789'){
-                                        echo "<th>更多選項</th>";
+                                    if(isset($_SESSION['mId'])){
+                                        if($_SESSION['mId']=='v123456789'){
+                                            echo "<th>更多選項</th>";
+                                        }
                                     }
                                 ?>
                         </tr>
@@ -152,17 +154,22 @@
                                         echo "<td>".$result[$i]['mDirector']."</td>";
                                         echo "<td>".$result[$i]['mPlot']."</td>";
                                         echo "<td><a href='movie_introduce.php?mTitle=$mTitle'>查看更多</a></td>";
-                                        if($_SESSION['mId']=='v123456789'){
-                                            $mTitle = $result [$i]['mTitle'];
-                                            echo "<td><a href='modify_movie.php?mTitle=$mTitle'>更新</a> / <a href='delete_movie.php?mTitle=$mTitle'>刪除</a></td>";
+                                        
+                                        if(isset($_SESSION['mId'])){
+                                            if($_SESSION['mId']=='v123456789'){
+                                                $mTitle = $result [$i]['mTitle'];
+                                                echo "<td><a href='modify_movie.php?mTitle=$mTitle'>更新</a> / <a href='delete_movie.php?mTitle=$mTitle'>刪除</a></td>";
+                                            }
                                         }
                                     echo "</tr>";
                                 }
                             ?>
                     </tbody>
                     <?php
-                        if($_SESSION['mId']=='v123456789'){
-                            echo "<a href='new_movie.php'>新增電影</a>";
+                        if(isset($_SESSION['mId'])){
+                            if($_SESSION['mId']=='v123456789'){
+                                echo "<a href='new_movie.php'>新增電影</a>";
+                            }
                         }
                     ?>
                 </table>
